+++ Source/Runtime/Public/Platform/Windows/PlatformWinTime.h
@@ -21,7 +21,7 @@ public:
     static void             Init();
     static void             Shutdown();
 
-    static float            Seconds();
+    static double           Seconds();
     static uint32_t         Milliseconds();
     static uint64_t         Microseconds();
